backend/requirements.txt
fastapi>=0.96.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
fastapi-users[sqlalchemy]>=12.0.0
alembic>=1.11.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.1
pytest-cov>=4.1.0
psycopg2-binary>=2.9.6
imap-tools>=1.0.0
defusedxml>=0.7.1
dynaconf>=3.1.12
python-dotenv>=1.0.0
tenacity>=8.2.2
apprise>=1.4.5
dnspython>=2.3.0
email-validator>=2.0.0
lxml>=4.9.2
zipfile36>=0.1.3
aiosmtplib>=2.0.2
jinja2>=3.1.2

# Gmail API (OAuth2)
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0